+++ packages/utilities/jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  preset: 'react-native',
-  testPathIgnorePatterns: ['<rootDir>/lib/'],
-};